Extract renderer config types
