+++ packages/react-cosmos/src/plugins/webpack/client/userDeps.ts
@@ -2,7 +2,7 @@ import {
   ReactFixturesByPath,
   ReactDecoratorsByPath
 } from 'react-cosmos-shared2/react';
-import { DomRendererConfig } from '../../../domRenderer';
+import { DomRendererConfig } from '../../../shared/rendererConfig';
 
 // NOTE: Renderer data is statically injected at compile time
 export const rendererConfig: DomRendererConfig = {
