+++ packages/react-cosmos/src/shared/rendererConfig.ts
@@ -1,0 +1,7 @@
+export type DomRendererConfig = {
+  containerQuerySelector: null | string;
+};
+
+export type NativeRendererConfig = {
+  port: number;
+};
